pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docvision"
version = "0.3.0"
description = "Production-ready document parsing with Vision Language Models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache License 2.0"}
authors = [
    {name = "Fahmi Aziz Fadhil", email = "fahmiazizfadhil09@gmail.com"},
]
keywords = ["vlm", "document-parsing", "ocr", "pdf", "vision-language-model", "agentic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=2.16.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.7",
    "opencv-python>=4.13.0",
    "numpy>=2.0.2",
    "scipy>=1.11.0",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "rapidocr-onnxruntime>=1.4.4",
    "huggingface-hub>=1.4.1",
    "tqdm>=4.67.2",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipykernel>=7.2.0",
]

[project.urls]
Homepage = "https://github.com/fahmiaziz98/doc-vision-parser"
Repository = "https://github.com/fahmiaziz98/doc-vision-parser"
Issues = "https://github.com/fahmiaziz98/doc-vision-parser/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/docvision"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests (e2e)",
]

[tool.coverage.run]
source = ["src/docvision"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
